+++ packages/offchain-resolver/package.json
@@ -30,8 +30,8 @@
     "docker:up": "docker compose up -d",
     "test": "yarn docker:up &&  yarn prisma:migrate && yarn mocha --require ts-node/register --extension .test.ts --recursive ./src",
     "start": "yarn prisma:migrate && yarn ts-node --transpile-only  ./src/index.ts",
-    "build": "yarn prisma generate --schema=./src/persistance/schema.prisma && yarn tsc",
-    "prisma:migrate": "yarn prisma migrate dev --name init --schema src/persistance/schema.prisma"
+    "build": "yarn prisma generate --schema=./src/persistence/schema.prisma && yarn tsc",
+    "prisma:migrate": "yarn prisma migrate dev --name init --schema src/persistence/schema.prisma"
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
