+++ packages/offchain-resolver/src/http/resolverEndpoint.test.ts
@@ -10,9 +10,9 @@ import { expect } from 'chai';
 import { ethers } from 'ethers';
 import express from 'express';
 import request from 'supertest';
-import { IDatabase } from '../persistance/IDatabase';
-import { clearDb } from '../persistance/clearDb';
-import { getDatabase, getDbClient } from '../persistance/getDatabase';
+import { IDatabase } from '../persistence/IDatabase';
+import { clearDb } from '../persistence/clearDb';
+import { getDatabase, getDbClient } from '../persistence/getDatabase';
 import { encodeEnsName } from './handleCcipRequest/dns/encodeEnsName';
 import { Interceptor } from './handleCcipRequest/handler/intercept';
 import { profile } from './profile';
